Renamed uploadControllerHelper to ImageRepository
